Realease 1.2.0 (#22)
* fix Readme.md

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update .travis.yml

* Update coverage badge

* Refactor degreesToDirection.

* Add getHeight() to return cloud height in feet

The noun "height" has been deliberately chosen since METARs report
cloud as height AGL rather than altitude AMSL.

* Allow cloud height to be set in feet

* Standardise cloud height internally

The noun should be height instead of altitude since cloud is reported as
AGL rather than AMSL. The calculations have also been swapped over so
the height is stored as it is reported in feet - the approximate
conversion to meters is still available via the setAltitude and
getAltitude methods for backwards compatibility.

* Change cloud parser to use setHeight

* Deprecate the cloud *Altitude methods

Since cloud is reported as height AGL, then the noun "altitude" is
incorrect and misleading.

* Update README.md

* Added Eclipse plugin ObjectAid for uml class diagram.

This plugin generates the class diagram from the java code and then
exports a picture of the diagram. For an unknown reason when exporting
the diagram, arrows on the relation disappear.

* hotfix/fixReadme: Fix version of pom

* Fix bug VariableWindParse: Update pom version.

Update pom version to 1.1.

* Fix regex of wind.

The regex used a \d in order to match the direction. But when the
direction wasn't in degrees but VRB the regex wasn't matched and the
windPart was not parsed. Changing \d to \w fix the problem.

* Change type of directionDegrees.

Change type from int to integer.
This change allows to have a null value for directionDegrees when the
direction is Variable.

* fix checkstyle

* fix checkstyle

* Updated pom version.

* Updated pom.

Added information on continuous integration and issue management.

* updated model

* Fix checkstyle in enum package.

* Fix checkstyle in facade.

* Fix checkstyle in builder and exception package.

* fix checkstyle in model package.

* Renamed WeatherCode to AbstractWeatherCode

* fix checkstyle in enums.

* fix checkstyles in facade.

* Fix checkstyles in utils package.

* Fix checkstyle in parser package

* updated model

* Added checkstyle file.

Added configuration file for codeclimate.

* Update minimum coverage

* update codeclimate.yml

* Fix spelling

* Added class abstractWeatherContainer.

It is a class containing:
- Wind
- Visibility
- list of clouds
- list of weather conditions
- vertical visibility

* Updated model.jpg

* Fix format

* Made AbstractWeatherContainer parent of AbstractWeatherCode

* added test for addWeatherCondition and addTest

* Feature/metar trends (#18)

* Added messages for trends.

* Added classes related to time in metar trends.

* Added hierarchy classes for trends.

AbstractTrend: the abstract super class. Contains wind, clouds,
weatherConditions and visibility.

AbstractMetarTrend: Subclass of AbstractTrend. Abstract class for metar
trends. Contains attributes for the time of the trend.

BECMGMetarTrend: Subclass of AbstractMetarTrend. Represents the BECMG
part.

TEMPOMetarTrend: Subclass of AbstractMetarTrend. Represents the Tempo part.

* Pull up constants FM, TEMPO and BECMG

* Added stringToTime method in converter.

* fix test when locale is not english

* Added Trends to metar.

* Updated parse method of metar in order to parse trends.

Added tests.

Refactor trend times (AT, FM, TL) inside a List instead of 3 attributes.

* Created new model for trends and updated readme.

* Updated pom.

* fix checkstyle

* update model

* Fix code climate issue.

* fix code climate issues

* Made AbstractTrend a subclass of AbstractWeatherContainer

* Update pom for release.