Refactor the TAFTrend hierarchy and parsing
TAFTrend classes are now `TafTrend`, `TafProbTrend` and `FMTafTrend`.
Created a Parser for each of the trend class.
A TafTrendParserFactory returns the corresponding parser depending on  the first word of the trend.
A new Parser hierarchy has been created. The new superclass is the AbstractWeatherContainerParser.
Created the TAFTrendFactory to set the specific WeatherChangeType of the TafTrend.