Fix parsing of weatherCondition
Improved the method parseWeatherCondition. The WeatherCondition is only added to the list of WeatherCondition if the whole token was ingested. This prevents false positive WeatherCondition to appear in the list.
The parsing order of the Phenomenon is also fixed, the phenomenon appear in the same order in the list as they do in the token.